Moderator list fixes
We were passing token instead of id, so were always coming up empty.

Hidden moderators weren't being handled; we now take the current
session_id and, if that user is an admin, we include all hidden mods; if
a hidden moderator, we include the moderator id itself (because current
Session needs to see itself in the returned moderator list in order to
activate moderation features in the client).