Embed compact_poll failures into response
If we encountered a failure (e.g. polling a room that no longer exists)
we'd return an overall failure for the entire compact_poll, but this
breaks things because Session never gets *any* responses and doesn't
know how to deal with a 401 response on the outer request.

This changes it to embed error responses into the compact_poll result,
which seems to make Session happy.