Fix padding handling to match what Session does
Session doesn't just pad with nulls, it also puts a 0x80 just before the
nulls to indicate the start of padding.

Strip it off as well (and re-apply during retrieval).

Also fixes the room view javascript to properly trim off the 0x80 byte
and padding, which fixes it failing on some messages.