convert-privkey.py -> sogs.key_convert
This is so that it can be invoked as python3 -msogs.key_convert without
requiring an extra script to be installed for deb postinst
auto-conversion of keys.

Also makes it take arguments for input/output, and to stop loading from
sogs.config (because this may be invoked before sogs.config is fully
available/configured).