Generate protobuf code with buster protoc
Buster is the oldest distribution we support, so seems fine as a
fallback for checkout code being able to run without needing to install
protobuf-compiler.

What we have now fails with an older protobuf python lib that tries to
load it, but the older one appears fine with newer protobuf libs.