Use better_profanity; add silent profanity filtering
better_profanity has a fairly extensive built-in bad word list and
handles lots of things like word permutations.

This also adds the ability to silently filter: if enabled (which it is,
by default) then profanity is inserted but not distributed, so the
poster gets a successful post (which requires we return an actual
message id), but the message goes nowhere.  (It is still in the
database, though, which means we have the ability to let admins see
filtered messages).

This commit also fixes bugs in our message handling:
- open group messages a protobuf Content, not DataMessage, and so we
  were parsing it wrong (and getting invalid utf-8 data errors).
- we were not stripping padding correctly, nor re-adding padding
  correctly (but the first bug masked the second one) and so were still
  storing full-padding messages.