sqlite transaction deadlock fixes
Make transactions use BEGIN IMMEDIATE: non-immediate transactions
introduce a race condition that causes query failures if you have
concurrency that does a SELECT followed by a modification within the
same transaction: if two occur at the same time, one will fail when it
tries to update the transaction to a write transaction.

In particular we hit this case occasionally because, when posting, we
first do a SELECT to determine whether the user should be rate limited.
Thus we end up in a failure state when we do something like:

    - thread 1 BEGIN
    - thread 1 SELECT number of recent posts -- begins a read tx
    - thread 2 BEGIN
    - thread 2 SELECT number of recent posts -- begins a read tx
    - thread 1 INSERT new post -- upgrades to a write tx
    - thread 2 INSERT new post -- attempts to upgrade, but fails

When we use an immediate transaction that doesn't happen: thread 2
just gets blocked until thread 1 finishes.

Make this change isn't a big performance hit: we're currently only using
db.tx() where we are writing anyway, so this involves the same number of
write transactions, we just don't avoid the error-prone transaction
upgrade.

There is, however, no way to begin an immediate transaction with a
SAVEPOINT (the implicitly outermost save point transaction is always
DEFERRED), so this also changes the outermost transaction to be a BEGIN
IMMEDIATE/COMMIT/ROLLBACK rather than a savepoint (savepoints are now
used only when nesting).

Currently the savepoint-only approach here isn't working properly: if a
rollback occurs then the connection remains in a permanent "database is
locked" state.