Add capabilities and capability checking
New endpoint /capabilities returns the capabilities of this pysogs; it
only supports a `required` query parameter that takes a comma-separated
list of required capabilities and returns an error response code if any
of them are not satisfied.  (This, in particular, will be useful for batch
sequence chaining to abort a batch sequence immediately if a required
capability isn't provided by putting the required capabilities call at
the beginning of the sequence).