Test and fix ban timeouts, and allow global timeouts
Adds more tests for ban timeouts, which revealed a permission futures
bug that wouldn't properly apply updates when `banned` was NULL.

Fixing this required splitting permission updates and bans into separate
tables, which makes things easier to manage.

Moreover this allows the new user_ban_futures table to accept a NULL
room so that we can use it to also have global ban timeouts, which is
implemented here.