Add docstring-to-markdown script
This will replace the api.yaml content, which is a pain to actually
convert to markdown that doesn't suck and is also a pain to work with.

This version reads docstrings from all registered endpoints, puts
together the markdown, groups it by blueprint, and spits it out.

Additional content gets loaded from docs/snippets/*.md.  For instance
the "rooms" section prefix is the content of `rooms.md`; supplemental
content for some single endpoint comes from `rooms.update_room.md` (if
it exists), etc.