Small refactor database & privkey initialization
Currently just loading the sogs module will create a ./sogs.db and
./key_ed25519 if they don't exists, which is often undesirable.  This
refactors the database and privkey initialization so that the automatic
creation only occurs when running as a uwsgi app.

For `python3 -msogs` usage a new `--initialize` option is added to allow
explicitly enabling the creation; without this new flag the script fails
instead if the database doesn't exist.