Fix migration errors
The message_metadata table was causing upgrade errors because sqlite
rename only applies to immediate views, but not views-of-views (and
message_metadata is a view of message_details which is a view of
messages), and so the rename would fail.

This fixes it by dropping the views and recreating them later in the
import.

Also do this on postgresql, because the same scenario in postgresql
works, but doesn't rename the output column in the view which would also
cause problems.