Accommodate v0.2.0 import over-incrementing message_sequence
v0.2.0 had a trigger that added an extra (unnecessary) increment to
message_sequence (which was then called `updates`).  It didn't hurt
anything, but if we import via such a version we breaks the import tests
because the message_sequence value will be one too high, so hack it back
down by one to make it work.