Rejigger CLI database init/upgrades
Move most of the database loading and initialization in the CLI tool
until after arg parsing so that if we are doing something like `--help`
we don't get the lag of running db migration checks.

This also adds --upgrade and --check-upgrades arguments to the CLI that
upgrade and check for upgrades, respectively; migrations are now never
automatically performed, and any other option fails if an upgrade is
required rather than silently upgrading.

This carries through the database code by adding a `check_only` option
to migration code to raise an exception rather than performing upgrades
when an upgrade is needed.

Also includes a DRYing out of recreating the user_permissions view which
was being done in both seqno_etc and room_accessible migrations; moved
it to a separate migration that runs later to recreate.