Build the file response manually
send_file() has optimizations that won't work through an onion request
(in particular, we can't extract the response body from such a
response, but we need that for onion requests).

This changes it to build a flask.Response that streams it via Python
file iteration instead, which is a bit less efficient, but allows
extracting the body where needed for onion request handling.  (It should
still stream, rather than slam everything into memory, when invoked
directly).