Preserve unicode filenames & fix tests
Changes the file upload to only strip bad characters from the filename
we use to generate the path on disk, but not the suggested filename we
store in the database and return to the user.

Also changes the test suite to interpret the random byte filename as
latin1 so that it ends up as utf-8 for flask.  Otherwise flask does the
latin1 conversion itself, and returns utf-8 back, but that breaks the
test (because what we get back has gone through an unexpected latin1 ->
utf-8 encoding).