Add sogs blinded key permission migration
Adds migration of unblinded user permissions to blinded user permissions
on first login.

This works by composing a list of unmigrated ids at sogs startup: we
track the positive alternative of the two possible blinded Ed keys.
When a user first authenticates and doesn't exist, we check to see if
their |kA| is in the needs_migration table and, if it is, we construct
the new `users` row by copying over everything from the old (unblinded)
`users` row, updating all permissions and futures referencing the old id
to instead reference the new id, and then remove global settings on the
old account.

Effectively this *transfers* all permissions from the old account to the
new one, so that we aren't left with duplicates moderators/bans/etc.
anywhere.

This commit includes blinding tests, and removes the old derived key
code (which was from an obsolete, draft blinded id design).