Make future permissions restore room defaults
If we change permission to write=false, then schedule a future that
restores write=true, we end up (after arriving at the future) with an
explicit write=true permission for the user.

This seems slightly wrong because the positive permission is actually
*more* permissive than what the user *probably* had which was just write
permission inherited from the room.  This won't matter hugely right
away, but could show up later if the room default_write value gets
changed (at which point this disciplined user would have write access).

This changes the future to set the override to null if the future would
be setting the room default as a permission.  So, in the described
situation, we'd set `write=null` in the override row rather than
`write=true`, and so the user would end up with write permissions via
the default, rather than via a specific permission grant.