Fix test suite failure edge case
This could *rarely* fail, because if it runs really close to the integer
timestamp ticking over (e.g. at sometime.999), we could end up with an
integer timestamp that ticks over and passes validation when it
shouldn't because the server processing code gets it at
(sometime+1).001.

Increase it by 1 to avoid the error.