Add optimized view for fetching room moderators
Using the user_permissions view for this is very slow: because it has to
check each user.moderator value in the various COALESCEs it ends up
doing a full scan of the `users` table.

This adds a new `room_moderators` view for fetching moderator lists that
is a couple of orders of magnitude faster (by constructing the query
differently).