Reactions ordering
This redesigns how reactions are stored so that we track a single row
for each reaction, then add referencing user_reactions rows for each
user reaction.

This uses slightly less space, but more importantly lets us use the
reaction primary key to determine the proper order of the reactions.

This then adds a new `"index"` field into the returned reactions which
contains a 0..N-1 sequence indicating which order the reactions are
meant to be displayed (that is: the order in which the first of the
given reaction was added to the message).