Return new seqno from reaction-modifying endpoints
This allows Session to properly avoid a race condition between adding a
reaction and getting a reaction update without the reaction applied yet
(e.g. because the poll result was generated before the reaction update).

With the seqno returned, clients can use it to detect such a case to
properly handle local display of not-yet-updated reactions applied to a
post.