Return info_updates from room-modifying endpoints
This works similarly to returning the seqno from post-modifying
endpoints, but for room metadata (fields, pinned messages, mods, etc.).

This allows a client to track when the expected change should have been
applied to the room when there might be concurrent room polling that
could race with a pre-update response.