Fix info_updates not updating on hidden global mod changes
The triggers managing info_updates where not updating when a hidden
global mod/admin was added, but that is wrong because mods/admins of the
room *do* see those entries and need to know (via a changed room
info_updates when polling for new room info) when the mod list changes,
even from a hidden global mod change.

This updates the trigger to apply regardless of the visible_mod value.