Check for blinding when de-modding/unbanning
This allows a Session client to remove a user by *unblinded* ID to have
the server convert that id to the blinded ID (when blinding is enabled).

Without this, you have to use the blinded id to unban or remove a
moderator.  (And this makes it work like the *adding* a mod or banning).