filter responses, overhaul filter configuration
This adds an ability to send responses to filtered messages either back
to (just) the user who sent them (via a whisper), or to the whole room.

This allows configuration of the message (or messages, for random
selection), profile name, and uses a (simple) templating mechanism for
the message so that you can (somewhat) personalize it for the user.

This builds off PR #129 for language filtering, but revamps how
languages are specified and overridden with a mechanism that we can use
in the future for other room-specific settings, and that doesn't rely on
room ids (but rather tokens).

Other changes included here:

- You can now control whether or not mods/admins are affected by
  filtering via a new `filter_mods` setting.

- profanity filtering can be room specific.

- alphabet filtering can be room specific (revamping how PR #129 did
  room-specific settings).

- alphabet filtering can be configurably silent or not, like profanity
  filtering.

- added a new, separate example .ini file showing how to set things up
  with room-specific settings.