Add seqno_creation to be able to skip irrelevant deletions
Add a new column messages.seqno_creation that lets us track a message's
initial seqno value.  This allows us to skip deletion tombstones for
messages that the user doesn't have (i.e. because the message was both
created and then deleted after the client's current seqno) which can
significantly drop the number of messages we have to deliver when
syncing up a room with a lot of deletions.