Fix room deletion when a room has reactions
Room deletion wasn't working because there isn't a cascade on the
foreign key from user_reactions -> reactions, and so deleting a room
with reactions present would fail.  This first removes all the
user_reactions to fix it.

(This lack of cascading delete here is intentional: typically, you
aren't supposed to delete from the reactions end, but rather from the
user_reactions end with removal of a `reaction` being automatically
triggered once there are no more referencing user_reactions).

Includes a test (which is confirmed broken without the fix included
here).