Just correct the nginx config proxy_pass pattern in comments
Notes:

When use the ` http://127.3.2.1:4242$request_uri ` in the nginx config
file `sogs-proxy`, it will redirect to error page `
http://127.3.2.1:4242/r/room_token ` by clicking the room link on the
SOGS index web page.

When use the ` http://127.3.2.1:4242/ ` instead, it will redirect to `
http://example.com/r/room_token/ ` or ` http://<public_ip>/r/room_token
`, and this works well for the SOGS web accessing.

I also commit this fix and make a PR to the deb package building branch
` debian/sid `.