Fix the nginx proxy_pass pattern
Notes:

When use the ` http://127.3.2.1:4242$request_uri ` in the nginx config
file `sogs-proxy`, it will redirect to a ERROR page `
http://127.3.2.1:4242/r/room_token ` by clicking the room link on the
SOGS index web page.

When use the ` http://127.3.2.1:4242/ ` instead, it will redirect to `
http://example.com/r/room_token/ ` or ` http://<public_ip>/r/room_token
`, and this works well for the SOGS index web accessing.

I just found, fixed and tested this issue recently!